Sketchup integration
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@1121 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ GPUKernel.h
@@ -390,7 +390,7 @@ protected:
 	int m_nbActiveMaterials;
    int m_nbActiveTextures;
    int m_lightInformationSize;
-   int m_maxPrimitivesPerBox;
+   size_t m_maxPrimitivesPerBox;
 	Vertex		 m_viewPos;
 	Vertex		 m_viewDir;
 	Vertex		 m_angles;
